Add core::dm_udev_sync sub-module and UdevSync type
Add a new sub-module to encapsulate creating, destroying, and
interacting with device-mapper Udev synchronisation semaphores.

The module contains the low-level wrappers to the unsafe SysV semaphore
bindings provided by the libc crate as well as higher level UdevSync
interface for use by other sub-modules:

  UdevSync::begin()
  UdevSync::end()
  UdevSync::cancel()

This provides the necessary primitives to implement the ioctl-issuing
side of the DM cookie protocol. Completion of DM cookie semaphores is
not implemented as this is always carried out by the dmsetup program via
udev rule callbacks.

This simplifies integrating udev synchronization with the other
DM methods and avoids the need to expose the low-level semaphore
operations outside the dm_udev_sync module.