Add core::dm_udev_sync
Add a new sub-module to encapsulate creating, destroying, and
interacting with device-mapper Udev synchronisation semaphores.

The module contains the low-level wrappers to the unsafe SysV semaphore
bindings provided by the libc crate as well as higher level functions
exported for use by other sub-modules.

This provides the necessary primitives to implement the ioctl-issuing
side of the DM cookie protocol. Completion of DM cookie semaphores is
not implemented as this is always carried out by the dmsetup program via
udev rule callbacks.