Work around udev failures in cachedev::tests::loop_test_minimal_cache_dev
Allowing udev time to interact with the devices invalidates some of the
assumptions in loop_test_minimal_cache_dev regarding the state of the
cache (since udev will issue reads to the device to determine its
content, meaning that the cache is no longer completely clean).

Temporarily disable the following assertions:

        assert_eq!(usage.used_cache, DataBlocks(0));
        assert_eq!(performance.read_hits, 0);
        assert_eq!(performance.read_misses, 0);
        assert_eq!(performance.promotions, 0);

This should be addressed by either suppressing udev rule processing for
this test, or by relaxing the assumptions made about the state of the
cache at the end of testing.