Work around udev race with xfs_admin -U <new uuid> in loop_test_basic_udev
The loop_test_basic_udev case creates an XFS file system, mounts and
umounts it, and then changes the file system UUID and tests for the
presence of the correct /dev/disk/by-uuid symlink.

This currently fails intermittently because the symlink is not present.

Introducing a brief sleep or a udev_settle() call *before* invoking
xfs_admin ensures that the symlink is present at the time of the check.

There is an existing udev_settle() call in verify() that should be
sufficient to ensure the symlink is properly set - it is unclear why the
event appears to be lost when xfs_admin is called immediately after the
umount.

It is not currently possible to reproduce this behaviour outside of the
test suite, for e.g. with a thin device set up at /dev/mapper/thin1:

mkfs.xfs -f /dev/mapper/thin1
mount /dev/mapper/thin1 /mnt
umount /dev/mapper/thin1
NEWUUID=$(uuidgen)
xfs_admin -U $NEWUUID /dev/mapper/thin1
udevadm settle
ls -l /dev/disk/by-uuid/$NEWUUID

Consistently passes without requiring the first udev settle command to
be un-commented.