Extend devicemapper-rs-sys bindings to include libdevmapper constants
Avoid manually re-defining the constants from libdevmapper (for e.g. the
udev flag values) by incorporating these definitions into the bindings
generated for devicemapper-rs-sys.

An allowlist is used to control the visibility of types and constants in
the bindings to prevent including unwanted interfaces from libdevmapper,
and to avoid a bindgen bug in the handling of the 'await' Rust keyword:

  https://github.com/rust-lang/rust-bindgen/pull/2247